requirements.txt
asgiref==3.10.0
attrs==25.4.0
black==25.9.0
click==8.3.0
colorama==0.4.6
dj-database-url==3.0.1
dj-email-url==1.0.6
Django==5.2.7
django-cache-url==3.4.5
djangorestframework==3.16.1
drf-spectacular==0.28.0
environs==14.3.0
flake8==7.3.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
marshmallow==4.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
python-dotenv==1.1.1
pytokens==0.2.0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.27.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
